+++ src/trx/game/stats/init.c
@@ -25,7 +25,7 @@
 
 #define M_FNV_1A_BASE 14695981039346656037ULL
 #define M_FNV_1A_PRIME 1099511628211ULL
-#define M_CACHE_VERSION 1
+#define M_CACHE_VERSION 3
 #define M_CACHE_FILENAME "max_stats.cache.json"
 
 static LEVEL_MAX_STATS *m_Stats = nullptr;
@@ -122,6 +122,11 @@ static JSON_OBJECT *M_SerializeLevelMaxStats(const LEVEL_MAX_STATS *const stats)
     JSON_ObjectAppendInt64(
         out, "max_kill_count", (int64_t)stats->max_kill_count);
     JSON_ObjectAppendInt64(
+        out, "max_kill_ally_count", (int64_t)stats->max_kill_ally_count);
+    JSON_ObjectAppendInt64(
+        out, "max_kill_non_ally_count",
+        (int64_t)stats->max_kill_non_ally_count);
+    JSON_ObjectAppendInt64(
         out, "max_pickup_count", (int64_t)stats->max_pickup_count);
     JSON_ObjectAppendInt64(
         out, "max_secret_count", (int64_t)stats->max_secret_count);
@@ -166,6 +171,10 @@ static bool M_DeserializeLevelMaxStats(
         obj, "max_pickup_secret_count", (int64_t)out->max_pickup_secret_count);
     out->max_kill_count = (size_t)JSON_ObjectGetInt64(
         obj, "max_kill_count", (int64_t)out->max_kill_count);
+    out->max_kill_ally_count = (size_t)JSON_ObjectGetInt64(
+        obj, "max_kill_ally_count", (int64_t)out->max_kill_ally_count);
+    out->max_kill_non_ally_count = (size_t)JSON_ObjectGetInt64(
+        obj, "max_kill_non_ally_count", (int64_t)out->max_kill_non_ally_count);
     out->max_pickup_count = (size_t)JSON_ObjectGetInt64(
         obj, "max_pickup_count", (int64_t)out->max_pickup_count);
     out->max_secret_count = (size_t)JSON_ObjectGetInt64(
@@ -403,9 +412,11 @@ void Stats_CalculateMaxStats(void)
         LOG_INFO(
             "Level %d (%s)", GF_GetLevelOrdinalNumber(GFLT_MAIN, level),
             level->title);
-        LOG_INFO("    pickups: %d", max_stats->max_pickup_count);
-        LOG_INFO("    kills:   %d", max_stats->max_kill_count);
-        LOG_INFO("    secrets: %d", max_stats->max_secret_count);
+        LOG_INFO("    pickups:   %d", max_stats->max_pickup_count);
+        LOG_INFO("    kills:     %d", max_stats->max_kill_count);
+        LOG_INFO("      allies:  %d", max_stats->max_kill_ally_count);
+        LOG_INFO("      enemies: %d", max_stats->max_kill_non_ally_count);
+        LOG_INFO("    secrets:   %d", max_stats->max_secret_count);
 #endif
     }
 
@@ -415,6 +426,8 @@ finish:
     const FINAL_STATS final_stats = Stats_ComputeFinalStats(true);
     LOG_INFO("Max pickups: %d", final_stats.max_stats.max_pickup_count);
     LOG_INFO("Max kills:   %d", final_stats.max_stats.max_kill_count);
+    LOG_INFO("  allies:    %d", final_stats.max_stats.max_kill_ally_count);
+    LOG_INFO("  enemies:   %d", final_stats.max_stats.max_kill_non_ally_count);
     LOG_INFO("Max secrets: %d", final_stats.max_stats.max_secret_count);
     Benchmark_End(&benchmark, nullptr);
 }