+++ docs/CHANGELOG.md
@@ -30,6 +30,7 @@
 - improved savegame file size by reducing it about 20–30%.
 - changed photo mode to no longer show "Entering photo mode" in the console
 - changed photo mode to always display a red frame around the game view when active (not visible in screenshots)
+- changed stats dialog to include allies in kill count if they turn hostile. This applies to all levels that follow, and the final stats screen.
 - changed rooms-to-draw tracking to no longer stop at the 100-room limit
 - changed boulders to stop if the ceiling height is lower than their height
 - changed all UI bar colors from hardcoded to configurable via `cfg/ui.json5`, enabling some customization for PS1 bars